Backend_Assignment/org-management/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.29.0
motor==3.5.1
pymongo==4.6.3
pydantic==2.9.2
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
loguru==0.7.2
pytest==8.2.0
pytest-asyncio==0.23.5
httpx==0.27.0
mongomock_motor==0.0.21
